backend/requirements.txt
asgiref==3.2.10
Django==2.2.19
django-colorfield==0.6.3
django-extra-fields==3.0.2
django-filter==21.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
gunicorn==20.0.4
oauthlib==3.2.0
Pillow==9.0.1
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
